requirements.txt
# --- Web Framework ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.4
python-dotenv>=1.0.1
python-multipart>=0.0.9

# --- LangChain ---
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.0
langchain-google-vertexai>=2.0.0

# --- Database ---
psycopg2-binary>=2.9.9
psycopg[binary,pool]>=3.1.18
SQLAlchemy>=2.0.30

# --- AWS ---
boto3>=1.34.110

# --- Docker ---
docker>=7.1.0

# --- SSH & Crypto ---
paramiko>=3.4.0
cryptography>=42.0.7

# --- Git ---
GitPython>=3.1.43

# --- GitHub API ---
PyGithub>=2.3.0

# --- Utilities ---
requests>=2.32.2